Refactor SecurityConfig and add Authentication and JWT configs
In order to improve the structure and maintainability of the codebase, SecurityConfig class has been refactored to separate authentication and JWT configurations. The specifics of the authentication and the JWT handling logic has been separated out into new classes named AuthenticationConfig and JwtConfig respectively. This refactoring puts similar functions together improving code readability and future adaptations while adhering to the single responsibility principle.
