+++ ecommerce-service/src/main/java/dev/ixale/ecommerceservice/config/AuthenticationConfig.java
@@ -1,0 +1,57 @@
+package dev.ixale.ecommerceservice.config;
+
+import dev.ixale.ecommerceservice.enums.Authority;
+import dev.ixale.ecommerceservice.service.UserService;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+//@Ena
+public class AuthenticationConfig {
+    private final UserService userService;
+
+    public AuthenticationConfig(UserService userService) {
+        this.userService = userService;
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    @Bean
+    public AuthenticationProvider authenticationProvider(PasswordEncoder passwordEncoder) {
+        DaoAuthenticationProvider provider = new DaoAuthenticationProvider();
+        provider.setPasswordEncoder(passwordEncoder);
+        provider.setUserDetailsService(userService);
+        return provider;
+    }
+
+    @Bean
+    public AuthenticationManager authenticationManager(HttpSecurity http, PasswordEncoder passwordEncoder) throws Exception {
+        // Gets the Authentication Manager Builder Class from the HttpSecurity Object
+        AuthenticationManagerBuilder authManagerBuilder =
+                http.getSharedObject(AuthenticationManagerBuilder.class);
+
+        authManagerBuilder.authenticationProvider(authenticationProvider(passwordEncoder));
+        authManagerBuilder.inMemoryAuthentication()
+                .withUser("admin")
+                .password(passwordEncoder.encode("admin"))
+                .roles("SU")
+                .authorities(Authority.ADMIN.toString(), Authority.WRITE.toString(), Authority.READ.toString());
+        authManagerBuilder.inMemoryAuthentication()
+                .withUser("user")
+                .password(passwordEncoder.encode("user"))
+                .authorities(Authority.USER.toString(), Authority.READ.toString());
+
+        return authManagerBuilder.build();
+    }
+
+}
